Completed city unit transfers

+++ Stratagem/Game Scenes/City/City.swift
@@ -30,7 +30,7 @@ public class City {
     /// Stats
     var resources: [ResourceTypes: CGFloat] = [.POPULATION: 1000, .CREDITS: 1000, .METAL: 50] // Firebase
     var resourcesCap: [ResourceTypes: CGFloat] = [.POPULATION: 3000, .CREDITS: 100000, .METAL: 500]
-    var units: [UnitType: CGFloat] =
+    var units: [UnitType: Int] =
         [.BRAWLER: 10, .SNIPER: 5, .FIGHTER: 30]
     
     /// Initializes city variables (required). If not terrain is provided, create a new city
